Update README.md
Laboratory work at ITMO University, programming. first year

Laboratory work #1, variant 312621

Write a Java program that performs the appropriate action options. The program must meet the following requirements:

It should be packaged into an executable jar archive.
The expression must be evaluated in accordance with the calculation of mathematical expressions (the order of actions must be observed, etc.).
The program must use math functions from the Java standard library.
The result of evaluating the expression to be output to standard output must be output in the specified format.
The program must be executed on the helios server.

The laboratory report should contain:

Task text.
Source code of the program.
The result of the program.
Conclusions on the work.